chore: SecurityConfigTest 수정
- yml에 redis 환경 변수 추가
- PasswordEncoderTest 추가
